feat(ui): implement new UI renderer with WebGPU support
- Added ui_renderer.cpp and ui_renderer.h for the new UI rendering system.
- Introduced LayerStorage for managing primitive and index data.
- Implemented texture and buffer creation functions for WebGPU.
- Added shader module creation and rendering pipeline setup.
- Integrated UI layer management and rendering logic.
- Updated ui_scroll_view.c to ensure proper float calculations for scroll bar positioning.
- Refactored ui_state.c to clear and reserve space for edit string.
- Changed edit_str type in ui_state.h from ustring_view to std::string for better memory management.
- Updated ui_type.h to include core/simd.h instead of foundation/simd.h.
- Modified union.h to include webgpu_context.h and core/api.h.
- Created webgpu_context.cpp and webgpu_context.h for managing WebGPU context and resources.
- Updated stb_image_write.h to use snprintf instead of sprintf for safer string formatting.

+++ src/ui/ui_scroll_view.c
@@ -38,7 +38,7 @@ void ui_scroll_view(ui_scroll_view_t *view, ui_rect rect, u32 layer_index, u32 c
     const f32 scroll_bar_height = MACRO_MIN(1.f, rect.h / content_height) * rect.h;
     ui_rect bar_rect = (ui_rect){
         .x = rect.x + rect.w - (hover || active ? UI_SCROLL_BAR_ACTIVE_WIDTH : UI_SCROLL_BAR_DEFAULT_WIDTH),
-        .y = rect.y + view->offset_y / content_height * rect.h,
+        .y = (f32)(rect.y + view->offset_y / content_height * rect.h),
         .w = UI_SCROLL_BAR_DEFAULT_WIDTH,
         .h = scroll_bar_height};
